# Core: Remove Location.__hash__ for faster generation